FUCK
WE ARE ASSHOLES we forgot to include jQuery in prefs.html and now that I added it it doesn't work.
We need to test the whole authLogin() in the server using login.html and rebuild the cookies authentication in prefs.html

we're fucking stupid.

Also, make sure it doesn't work if you don't have your jQuery somehow (LIKE WE FUCKING DID).